+++ graph.cuh
@@ -6,11 +6,11 @@
 #include <cstdlib>
 
 struct Graph {
-    std::vector<int> adjacencyList; // all edges
-    std::vector<int> edgesOffset; // offset to adjacencyList for every vertex
-    std::vector<int> edgesSize; //number of edges for every vertex
-    int numVertices = 0;
-    int numEdges = 0;
+    std::vector<long int> adjacencyList; // all edges
+    std::vector<long int> edgesOffset; // offset to adjacencyList for every vertex
+    std::vector<long int> edgesSize; //number of edges for every vertex
+    long int numVertices = 0;
+    long int numEdges = 0;
 };
 
 void readGraph(Graph& G);
